refactor channels dialog out of pdsview and use dstack
